change: Add name space `message_handler` and rename message_handler to core_handler
